handle test diffs from merge
