+++ rust/util_libs/mocks/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod host;
 pub mod mongodb_runner;
 pub mod nats_message;
+pub mod test_service_response;
 pub mod workload;
